+++ packages/error-handler/src/index.ts
@@ -1,6 +1,7 @@
 //errors
 export * from "./errors/HttpError.js";
 export * from "./errors/CustomError.js";
+export * from "./errors/InputValidationError.js";
 export * from "./errors/types.js";
 
 //error handler middle