+++ packages/auth/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @khaled/auth
 
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @khaled/error-handler@1.2.0
+
 ## 1.1.0
 
 ### Minor Changes