+++ packages/error-handler/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @khaled/error-handler
 
+## 1.2.0
+
+### Minor Changes
+
+- support inputvalidation error and use zod error with it
+
 ## 1.1.1
 
 ### Patch Changes